backend/requirements/base.txt
Django>=5.2,<5.3
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.5,<6.0
django-cors-headers>=4.9,<5.0
django-filter>=25.2,<26.0
drf-spectacular>=0.29,<1.0
psycopg2-binary>=2.9,<3.0
Pillow>=12.0,<13.0
python-decouple>=3.8,<4.0
